+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/DisallowedTypeConverter.cs
@@ -3,14 +3,12 @@
 
 namespace System.Text.Json.Serialization.Converters
 {
-    internal sealed class UnsupportedTypeConverter<T> : JsonConverter<T>
+    internal sealed class DisallowedTypeConverter<T> : JsonConverter<T>
     {
-        private static readonly string s_exceptionMessage = SR.Format(SR.SerializeTypeInstanceNotSupported, typeof(T).FullName);
-
         public override T Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) =>
-            throw new NotSupportedException(s_exceptionMessage);
+            throw new NotSupportedException(SR.Format(SR.SerializeTypeInstanceNotSupported, typeof(T).FullName));
 
         public override void Write(Utf8JsonWriter writer, T value, JsonSerializerOptions options) =>
-            throw new NotSupportedException(s_exceptionMessage);
+            throw new NotSupportedException(SR.Format(SR.SerializeTypeInstanceNotSupported, typeof(T).FullName));
     }
 }
