--- src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Value/TypeConverter.cs
@@ -1,18 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System.Text.Json.Serialization.Converters
-{
-    internal sealed class TypeConverter : JsonConverter<Type>
-    {
-        public override Type Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            throw new NotSupportedException(SR.SerializeTypeInstanceNotSupported);
-        }
-
-        public override void Write(Utf8JsonWriter writer, Type value, JsonSerializerOptions options)
-        {
-            throw new NotSupportedException(SR.SerializeTypeInstanceNotSupported);
-        }
-    }
-}
